# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CUP FICCT Backend API",
        "description": "API REST del sistema de admisión al CUP FICCT.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor del backend"
        }
    ],
    "paths": {
...
        "/api/applicant-admission/postulantes/{postulante}/postulaciones/{convocatoria}/verificar": {
            "put": {
                "tags": [
                    "Postulaciones"
                ],
                "summary": "Verificar requisitos (aprueba la inscripción al CUP)",
                "operationId": "7c659b54f0ea23582d23b4705b08897e",
                "parameters": [
                    {
                        "name": "postulante",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "convocatoria",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Postulación VERIFICADO",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Postulacion"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applicant-admission/postulantes/{postulante}/postulaciones/{convocatoria}/rechazar": {
            "put": {
                "tags": [
                    "Postulaciones"
                ],
                "summary": "Rechazar por no cumplir requisitos",
                "operationId": "6aef490d65a5cf4bb8c82cbc98d8b4e3",
                "parameters": [
                    {
                        "name": "postulante",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "convocatoria",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "motivo"
                                ],
                                "properties": {
                                    "motivo": {
                                        "type": "string",
                                        "example": "Falta certificado de nacimiento"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Postulación RECHAZADO",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Postulacion"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Postulacion": {
                "properties": {
                    "postulante_documento": {
                        "type": "string",
                        "example": "9876543"
                    },
                    "convocatoria_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "carrera_primera_codigo": {
                        "type": "string",
                        "example": "187-09"
                    },
                    "carrera_segunda_codigo": {
                        "type": "string",
                        "example": "187-10"
                    },
                    "estado": {
                        "type": "string",
                        "example": "PENDIENTE",
                        "enum": [
                            "PENDIENTE",
                            "VERIFICADO",
                            "RECHAZADO"
                        ]
                    },
                    "observacion": {
                        "type": "string",
                        "example": "Falta certificado de nacimiento",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}